requirements.txt
kafka-python
flask
scikit-learn
pandas
pyyaml
